+++ MelonLoader/Preferences/Category.cs
@@ -54,11 +54,7 @@ namespace MelonLoader
             entry.DisplayName = displayname;
             entry.Hidden = hidden;
             entry.DefaultValue_bool = value;
-            if (entry.Type != MelonPreferences_Entry.TypeEnum.BOOL)
-            {
-                entry.Type = MelonPreferences_Entry.TypeEnum.BOOL;
-                entry.ValueEdited_bool = entry.Value_bool = value;
-            }
+            entry.ConvertCurrentValueType(MelonPreferences_Entry.TypeEnum.BOOL);
             return entry;
         }
         public MelonPreferences_Entry CreateEntry(string name, int value, string displayname = null, bool hidden = false)
@@ -71,11 +67,7 @@ namespace MelonLoader
             entry.DisplayName = displayname;
             entry.Hidden = hidden;
             entry.DefaultValue_int = value;
-            if (entry.Type != MelonPreferences_Entry.TypeEnum.INT)
-            {
-                entry.Type = MelonPreferences_Entry.TypeEnum.INT;
-                entry.ValueEdited_int = entry.Value_int = value;
-            }
+            entry.ConvertCurrentValueType(MelonPreferences_Entry.TypeEnum.INT);
             return entry;
         }
         public MelonPreferences_Entry CreateEntry(string name, float value, string displayname = null, bool hidden = false)
@@ -88,11 +80,7 @@ namespace MelonLoader
             entry.DisplayName = displayname;
             entry.Hidden = hidden;
             entry.DefaultValue_float = value;
-            if (entry.Type != MelonPreferences_Entry.TypeEnum.FLOAT)
-            {
-                entry.Type = MelonPreferences_Entry.TypeEnum.FLOAT;
-                entry.ValueEdited_float = entry.Value_float = value;
-            }
+            entry.ConvertCurrentValueType(MelonPreferences_Entry.TypeEnum.FLOAT);
             return entry;
         }
         public MelonPreferences_Entry CreateEntry(string name, long value, string displayname = null, bool hidden = false)
@@ -105,11 +93,7 @@ namespace MelonLoader
             entry.DisplayName = displayname;
             entry.Hidden = hidden;
             entry.DefaultValue_long = value;
-            if (entry.Type != MelonPreferences_Entry.TypeEnum.LONG)
-            {
-                entry.Type = MelonPreferences_Entry.TypeEnum.LONG;
-                entry.ValueEdited_long = entry.Value_long = value;
-            }
+            entry.ConvertCurrentValueType(MelonPreferences_Entry.TypeEnum.LONG);
             return entry;
         }
         public MelonPreferences_Entry CreateEntry(string name, double value, string displayname = null, bool hidden = false)
@@ -122,74 +106,7 @@ namespace MelonLoader
             entry.DisplayName = displayname;
             entry.Hidden = hidden;
             entry.DefaultValue_double = value;
-            if (entry.Type != MelonPreferences_Entry.TypeEnum.DOUBLE)
-            {
-                entry.Type = MelonPreferences_Entry.TypeEnum.DOUBLE;
-                entry.ValueEdited_double = entry.Value_double = value;
-            }
-            return entry;
-        }
-
-        internal MelonPreferences_Entry LoadEntry(string name, string value, string displayname = null, bool hidden = false)
-        {
-            if (string.IsNullOrEmpty(name))
-                throw new Exception("Name is null or empty when calling LoadEntry");
-            MelonPreferences_Entry entry = GetEntry(name);
-            if (entry == null)
-                return new MelonPreferences_Entry(this, name, value, displayname, hidden);
-            entry.ValueEdited_string = entry.Value_string = value;
-            return entry;
-        }
-        internal MelonPreferences_Entry LoadEntry(string name, bool value, string displayname = null, bool hidden = false)
-        {
-            if (string.IsNullOrEmpty(name))
-                throw new Exception("Name is null or empty when calling LoadEntry");
-            MelonPreferences_Entry entry = GetEntry(name);
-            if (entry == null)
-                return new MelonPreferences_Entry(this, name, value, displayname, hidden);
-            entry.ValueEdited_bool = entry.Value_bool = value;
-            return entry;
-            
-        }
-        internal MelonPreferences_Entry LoadEntry(string name, int value, string displayname = null, bool hidden = false)
-        {
-            if (string.IsNullOrEmpty(name))
-                throw new Exception("Name is null or empty when calling LoadEntry");
-            MelonPreferences_Entry entry = GetEntry(name);
-            if (entry == null)
-                return new MelonPreferences_Entry(this, name, value, displayname, hidden);
-            entry.ValueEdited_int = entry.Value_int = value;
-            return entry;
-            
-        }
-        internal MelonPreferences_Entry LoadEntry(string name, float value, string displayname = null, bool hidden = false)
-        {
-            if (string.IsNullOrEmpty(name))
-                throw new Exception("Name is null or empty when calling LoadEntry");
-            MelonPreferences_Entry entry = GetEntry(name);
-            if (entry == null)
-                return new MelonPreferences_Entry(this, name, value, displayname, hidden);
-            entry.ValueEdited_float = entry.Value_float = value;
-            return entry;
-        }
-        internal MelonPreferences_Entry LoadEntry(string name, long value, string displayname = null, bool hidden = false)
-        {
-            if (string.IsNullOrEmpty(name))
-                throw new Exception("Name is null or empty when calling LoadEntry");
-            MelonPreferences_Entry entry = GetEntry(name);
-            if (entry == null)
-                return new MelonPreferences_Entry(this, name, value, displayname, hidden);
-            entry.ValueEdited_long = entry.Value_long = value;
-            return entry;
-        }
-        internal MelonPreferences_Entry LoadEntry(string name, double value, string displayname = null, bool hidden = false)
-        {
-            if (string.IsNullOrEmpty(name))
-                throw new Exception("Name is null or empty when calling LoadEntry");
-            MelonPreferences_Entry entry = GetEntry(name);
-            if (entry == null)
-                return new MelonPreferences_Entry(this, name, value, displayname, hidden);
-            entry.ValueEdited_double = entry.Value_double = value;
+            entry.ConvertCurrentValueType(MelonPreferences_Entry.TypeEnum.DOUBLE);
             return entry;
         }
     }